requirements.txt
torch
torch_geometric
opendssdirect-py
networkx
numpy